+++ src/store/strongstore/shardclient.h
@@ -81,7 +81,7 @@ namespace strongstore
 
     typedef std::function<void(int, const std::string &, const std::vector<std::pair<uint64_t, uint32_t>> &)> op_callback;
     typedef std::function<void(int, const std::string &, const std::vector<std::pair<uint64_t, uint32_t>> &)> op_timeout_callback;
-    typedef std::function<void(uint64_t, request_utils::Value, int, const std::vector<std::pair<uint64_t, uint32_t>> &)> transformed_callback;
+    typedef std::function<void(uint64_t, request_utils::Value, uint64_t, const std::vector<std::pair<uint64_t, uint32_t>> &)> transformed_callback;
 
     typedef std::function<void(int, Timestamp)> prepare_callback;
     typedef std::function<void(int, Timestamp)> prepare_timeout_callback;
